+++ internal/repository/recurring_expense_repository.go
@@ -14,7 +14,7 @@ var errRecurringExpenseNil error = errors.New("recurring expense is nil")
 type RecurringExpenseRepository interface {
 	List() ([]models.RecurringExpense, error)
 	GetByID(id uint) (*models.RecurringExpense, error)
-	GetByUserID(userID int) ([]models.RecurringExpense, error)
+	GetByUserID(userID uint) ([]models.RecurringExpense, error)
 	GetActiveByNextDate(nextDate time.Time) ([]models.RecurringExpense, error)
 	Create(recurringExpense *models.RecurringExpense) error
 	Update(recurringExpense *models.RecurringExpense) error
@@ -62,16 +62,16 @@ func (r *gormRecurringExpenseRepository) GetByID(id uint) (*models.RecurringExpe
 	return &recurringExpense, nil
 }
 
-func (r *gormRecurringExpenseRepository) GetByUserID(userID int) ([]models.RecurringExpense, error) {
+func (r *gormRecurringExpenseRepository) GetByUserID(userID uint) ([]models.RecurringExpense, error) {
 	r.logger.Debug("repo.recurring_expense.get_by_user_id",
 		slog.String("op", "repo.recurring_expense.get_by_user_id"),
-		slog.Int("user_id", userID),
+		slog.Uint64("user_id", uint64(userID)),
 	)
 	var recurringExpenses []models.RecurringExpense
 	if err := r.db.Preload("Category").Where("user_id = ?", userID).Order("next_date ASC").Find(&recurringExpenses).Error; err != nil {
 		r.logger.Error("repo.recurring_expense.get_by_user_id failed",
 			slog.String("op", "repo.recurring_expense.get_by_user_id"),
-			slog.Int("user_id", userID),
+			slog.Uint64("user_id", uint64(userID)),
 			slog.String("error", err.Error()),
 		)
 		return nil, err
